+++ R/PivotCalculation.R
@@ -19,8 +19,10 @@
 #'   typically used as base values for other calculations.
 #' @field displayOrder The order the calculations are displayed in the pivot
 #'   table.
-#' @field filters Any data filters specific to this calculation.  A PivotFilters
-#'   object.
+#' @field filters Any additional data filters specific to this calculation.
+#'   This can be a PivotFilters object that further restrict the data for the
+#'   calculation of a list of individual PivotFilter objects that provide more
+#'   flexability (and/or/replace).  See the Calculations vignette for details.
 #' @field format A character, list or custom function to format the calculation
 #'   result.
 #' @field dataName Specifies which data frame in the pivot table is used for
@@ -66,7 +68,7 @@ PivotCalculation <- R6::R6Class("PivotCalculation",
        checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", caption, missing(caption), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
        checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", visible, missing(visible), allowMissing=TRUE, allowNull=TRUE, allowedClasses="logical")
        checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", displayOrder, missing(displayOrder), allowMissing=TRUE, allowNull=TRUE, allowedClasses=c("integer", "numeric"))
-       checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", filters, missing(filters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
+       checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", filters, missing(filters), allowMissing=TRUE, allowNull=TRUE, allowedClasses=c("PivotFilters", "PivotFilterOverrides"))
        checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", format, missing(format), allowMissing=TRUE, allowNull=TRUE, allowedClasses=c("character","list","function"))
        checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", dataName, missing(dataName), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
        checkArgument(parentPivot$argumentCheckMode, FALSE, "PivotCalculation", "initialize", type, missing(type), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character", allowedValues=c("value", "summary", "calculation", "function"))
@@ -84,10 +86,7 @@ PivotCalculation <- R6::R6Class("PivotCalculation",
      }
      private$p_parentPivot <- parentPivot
      fstr <- NULL
-     if(!is.null(filters)) {
-       if (!("PivotFilters" %in% class(filters))) stop("PivotCalculation$new(): filters must be of type PivotFilters", call. = FALSE)
-       fstr <- filters$asString()
-     }
+     if(!is.null(filters)) fstr <- filters$asString()
      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculation$new", "Creating new Pivot Calculation...",
                                                                              list(calculationName=calculationName, caption=caption, visible=visible,
                                                                              displayOrder=displayOrder, filters=fstr, format=format, dataName=dataName,