+++ R/PivotCalculator.R
@@ -56,17 +56,11 @@
 #'   \item{\code{formatValue(value, format)}}{Formats a numerical value using
 #'   either an sprintf string, a list of arguments for the base::format()
 #'   function or using a custom R function.}
-#'   \item{\code{getFiltersForSingleValue = function(rowColFilters=NULL,
-#'   calcFilters=NULL)}}{Get the working filters for a single value
-#'   calculation.}
-#'   \item{\code{getFiltersForSummariseExpression = function(rowColFilters=NULL,
-#'   calcFilters=NULL)}}{Get the working filters for a summary calculation.}
-#'   \item{\code{getFiltersForCalculateFunction = function(rowColFilters=NULL,
-#'   calcFilters=NULL)}}{Get the working filters for a calculation based on
-#'   other calculations.}
+#'   \item{\code{getCombinedFilters = function(rowColFilters=NULL,
+#'   calcFilters=NULL)}}{Get the working filters for a calculation.}
 #'   \item{\code{getFiltersForNamedCalculation = function(calculationName=NULL,
 #'   calculationGroupName=NULL, rowColFilters=NULL, cell=NULL)}}{Get the working
-#'   filters for a custom calculation.}
+#'   filters for a named calculation.}
 #'   \item{\code{setWorkingData = function(cell=NULL)}}{Set the working
 #'   filters for a cell.}
 #'   \item{\code{evaluateSingleValue(dataFrame, workingFilters, valueName,
@@ -178,10 +172,9 @@ PivotCalculator <- R6::R6Class("PivotCalculator",
      if(private$p_parentPivot$argumentCheckMode > 0) {
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilters", filters1, missing(filters1), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilters")
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilters", filters2, missing(filters2), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilters")
+       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilters", action, missing(action), allowMissing=FALSE, allowNull=FALSE, allowedClasses="character", allowedValues=c("and", "replace", "or"))
      }
      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$setFilters", "Setting filters...")
-     checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilters", filters1, missing(filters1), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilters")
-     checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilters", filters2, missing(filters2), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilters")
      copy <- filters1$getCopy() # always copy, to avoid inadvertant bugs of a change to one filter affecting multiple cells in the Pivot Table
      copy$setFilters(filters=filters2, action=action)
      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$setFilters", "Set filters.")
@@ -191,6 +184,7 @@ PivotCalculator <- R6::R6Class("PivotCalculator",
      if(private$p_parentPivot$argumentCheckMode > 0) {
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilter", filters, missing(filters), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilters")
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilter", filter, missing(filter), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilter")
+       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilter", action, missing(action), allowMissing=FALSE, allowNull=FALSE, allowedClasses="character", allowedValues=c("and", "replace", "or"))
      }
      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$setFilter", "Setting filter...")
      copy <- filters$getCopy() # always copy, to avoid inadvertant bugs of changing one filter affecting multiple cells
@@ -203,6 +197,7 @@ PivotCalculator <- R6::R6Class("PivotCalculator",
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilterValues", filters, missing(filters), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotFilters")
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilterValues", variableName, missing(variableName), allowMissing=FALSE, allowNull=FALSE, allowedClasses="character")
        checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilterValues", values, missing(values), allowMissing=TRUE, allowNull=TRUE, mustBeAtomic=TRUE)
+       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "setFilterValues", action, missing(action), allowMissing=FALSE, allowNull=FALSE, allowedClasses="character", allowedValues=c("and", "replace", "or"))
      }
      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$setFilterValues", "Setting filter values...")
      copy <- filters$getCopy() # always copy, to avoid inadvertant bugs of changing one filter affecting multiple cells
@@ -318,67 +313,36 @@ PivotCalculator <- R6::R6Class("PivotCalculator",
      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$formatValue", "Formated value.")
      return(invisible(value))
    },
-   getFiltersForSingleValue = function(rowColFilters=NULL, calcFilters=NULL) {
+   getCombinedFilters = function(rowColFilters=NULL, calcFilters=NULL) {
      if(private$p_parentPivot$argumentCheckMode > 0) {
-       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getFiltersForSingleValue", rowColFilters, missing(rowColFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
-       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getFiltersForSingleValue", calcFilters, missing(calcFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
+       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getCombinedFilters", rowColFilters, missing(rowColFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
+       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getCombinedFilters", calcFilters, missing(calcFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses=c("PivotFilters", "PivotFilterOverrides"))
      }
-     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getFiltersForSingleValue", "Getting filters for single value...")
+     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getCombinedFilters", "Getting filters for calculation...")
      # get the final filter context for this calculation (calcFilters override row/col filters)
+     # see the comments above PivotFilter$setFilters() for rules/details on how the filters are combined
      rf <- list()
      rf$calculationFilters <- calcFilters
      filters <- NULL
      if(is.null(rowColFilters)) {
-       if(!isnull(calcFilters)) filters <- calcFilters$getCopy()
+       if(!isnull(calcFilters)) {
+         if("PivotFilters" %in% class(calcFilters)) filters <- calcFilters$getCopy()
+         else if("PivotFilterOverrides" %in% class(calcFilters)) {
+           filters <- PivotFilters$new(private$p_parentPivot)
+           calcFilters$apply(filters)
+           if(filters$isALL) filters <- NULL #i.e. if none of the calculation filters amounted to anything, then skip keeping the filters
+         }
+       }
      }
      else {
        filters <- rowColFilters$getCopy()
-       if(!is.null(calcFilters)) filters$setFilters(filters=calcFilters, action="and")
+       if(!is.null(calcFilters)) {
+         if("PivotFilters" %in% class(calcFilters)) filters$setFilters(filters=calcFilters, action="and")
+         else if("PivotFilterOverrides" %in% class(calcFilters)) calcFilters$apply(filters)
+       }
      }
      rf$workingFilters <- filters
-     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getFiltersForSingleValue", "Got filters for single value.")
-     return(invisible(rf))
-   },
-   getFiltersForSummariseExpression = function(rowColFilters=NULL, calcFilters=NULL) {
-     if(private$p_parentPivot$argumentCheckMode > 0) {
-       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getFiltersForSummariseExpression", rowColFilters, missing(rowColFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
-       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getFiltersForSummariseExpression", calcFilters, missing(calcFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
-     }
-     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getFiltersForSummariseExpression", "Getting filters for summary expression...")
-     # get the final filter context for this calculation (calcFilters override row/col filters)
-     rf <- list()
-     rf$calculationFilters <- calcFilters
-     filters <- NULL
-     if(is.null(rowColFilters)) {
-       if(!isnull(calcFilters)) filters <- calcFilters$getCopy()
-     }
-     else {
-       filters <- rowColFilters$getCopy()
-       if(!is.null(calcFilters)) filters$setFilters(filters=calcFilters, action="and")
-     }
-     rf$workingFilters <- filters
-     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getFiltersForSummariseExpression", "Got filters for summary expression.")
-     return(invisible(rf))
-   },
-   getFiltersForCalculateFunction = function(rowColFilters=NULL, calcFilters=NULL) {
-     if(private$p_parentPivot$argumentCheckMode > 0) {
-       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getFiltersForCalculateFunction", rowColFilters, missing(rowColFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
-       checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotCalculator", "getFiltersForCalculateFunction", calcFilters, missing(calcFilters), allowMissing=TRUE, allowNull=TRUE, allowedClasses="PivotFilters")
-     }
-     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getFiltersForCalculateFunction", "Getting filters for calculation function...")
-     # get the final filter context for this calculation (calcFilters override row/col filters)
-     rf <- list()
-     rf$calculationFilters <- calcFilters
-     filters <- NULL
-     if(is.null(rowColFilters)) {
-       if(!isnull(calcFilters)) filters <- calcFilters$getCopy()
-     }
-     else {
-       filters <- rowColFilters$getCopy()
-       if(!is.null(calcFilters)) filters$setFilters(filters=calcFilters, action="and")
-     }
-     rf$workingFilters <- filters
-     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getFiltersForCalculateFunction", "Got filters for calculation function.")
+     if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotCalculator$getCombinedFilters", "Got filters for calculation....")
      return(invisible(rf))
    },
    getFiltersForNamedCalculation = function(calculationName=NULL, calculationGroupName=NULL, rowColFilters=NULL, cell=NULL) {
@@ -405,16 +369,16 @@ PivotCalculator <- R6::R6Class("PivotCalculator",
          if((!is.null(cell))&&(cell$isTotal==TRUE)) {
            if(is.null(calc$summariseExpression)) { rf <- list() }
            else {
-             rf <- self$getFiltersForSummariseExpression(rowColFilters=rowColFilters, calcFilters=calc$filters)
+             rf <- self$getCombinedFilters(rowColFilters=rowColFilters, calcFilters=calc$filters)
            }
          }
          else {
-           rf <- self$getFiltersForSingleValue(rowColFilters=rowColFilters, calcFilters=calc$filters)
+           rf <- self$getCombinedFilters(rowColFilters=rowColFilters, calcFilters=calc$filters)
          }
          filters[[calc$calculationName]] <- rf
        }
        else if(calc$type=="summary") {
-         rf <- self$getFiltersForSummariseExpression(rowColFilters=rowColFilters, calcFilters=calc$filters)
+         rf <- self$getCombinedFilters(rowColFilters=rowColFilters, calcFilters=calc$filters)
          filters[[calc$calculationName]] <- rf
        }
        else if(calc$type=="calculation") {
@@ -422,7 +386,7 @@ PivotCalculator <- R6::R6Class("PivotCalculator",
          filters[[calc$calculationName]] <- rf
        }
        else if(calc$type=="function") {
-         rf <- self$getFiltersForCalculateFunction(rowColFilters=rowColFilters, calcFilters=calc$filters)
+         rf <- self$getCombinedFilters(rowColFilters=rowColFilters, calcFilters=calc$filters)
          filters[[calc$calculationName]] <- rf
        }
        else stop(paste0("PivotCalculator$getFiltersForNamedCalculation():  Unknown calculation type encountered '", calc$type,